\docType{data}
\name{mixtureA}
\alias{mixtureA}
\title{RnaQuin mixture A}
\format{Data frame for each RnaQuin sequin:
\itemize{
  \item Length: Length in nucleotide bases
  \item MXA: Input concentration in attomol/ul
}}
\usage{
data(mixtureA)
}
\description{
Sequin concentration for RnqQuin mixture A.
}