\docType{data}
\name{mixtureB}
\alias{mixtureB}
\title{RnaQuin mixture B}
\format{Data frame for each RnaQuin sequin:
\itemize{
  \item Length: Length in nucleotide bases
  \item MXB: Input concentration in attomol/ul
}}
\usage{
data(mixtureB)
}
\description{
Sequin concentration for RnqQuin mixture A.
}