\name{pltoAlignRead}
\alias{pltoAlignRead}
\title{Create alignment-reads plot}
\description{
Create an alignment-read plot where the x-axis is the input concentration and the y-axis is the measured number of reads aligned for each sequin.
}
\usage{
\method{pltoAlignRead}{TransQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{pltoAlignRead}{VarQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{pltoAlignRead}{FusQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{pltoAlignRead}{LadQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{pltoAlignRead}{MetaQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have input concentration (\code{input}) and measured abundance (\code{measured}).}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
  \item{showLOQ}{Show limit of alignment?}
}
\details{
Sequins are combined together across a range of concentrations to formulate a mixture. The mixture provides an internal scale against which to measure many features of an next-generation sequencing experiment, including read alignment in the accompying sample. The plot tabulates the input concentration against aligned reads, and evaluates the accuracy with a simple linear regression.

We would like to achieve a perfect correlation. However, this is typically not possible; the more corrected (slope and R2) the variables are, the better experiment it is.

Limit-of-alignment (LOA) is estimated if possible by piecewise linear segmentation.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=log2(data$input),
                  measured=log2(data$measured)
plotAlignRead(data)
}