\name{plotAllele}
\alias{plotAllele}
\title{Create an allele plot}
\description{
Create an allele plot where the x-axis is the expected allele frequency and the y-axis is the measured allele frequency.
}
\usage{
\method{plotAllele}{default}(data, showStats='left', title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}. The data set needs to have the expected allele frequency and measured allele frequency.}
  \item{showStats}{Where to show the regression equation? \code{\link{'right'}} or \code{\link{'left'}}. }
  \item{title}{Label of the plot.}
  \item{xname}{Label for the x-axis}
  \item{yname}{Label for the x-axis}
  \item{showLOQ}{Show limit of quantification by piecewise-linear segmentation?}
}
\details{
VarQuin are pairs of sequins that alternatively represent the reference and variant alleles with respect to the synthetic chromosome. By modulating the relative abundance of both the reference and the variant sequin, VarQuin enables a range of homozygous, heterozygous or heterogeneous allele frequencies to be emulated. For example, the use of only a single sequin can emulate a homozygous genotype, whilst the use of two sequins in equal parts can represent a heterozygous genotype.

The allele frequency plot tabulates the expected allele frequency defined by the sequins against measured allele frequency, and evaluates the accuracy with a simple linear regression. This would give us the sensitivity for a dynamic range of allele frations.

We would like to achieve a perfect correlation. However, this is typically not possible; instead we should use the regression to quantify the samples. The more corrected (slope and R2) the samples with the input concentration, the better experiment it is.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(VarQuin)
plotAllele(data)
}