\name{plotConjoint}
\alias{plotConjoint}
\title{Create conjoint plots}
\description{
Create scatter plot for conjoint sequins.
}
\usage{
plotConjoint(seqs, units, x, y, title=NULL, xlab=NULL, ylab=NULL)
}
\arguments{
  \item{seqs}{Sequin names}
  \item{units}{Copy units}
  \item{x}{Expected copy number on the x-axis}
  \item{y}{Measued abundance on the y-axis}
  \item{title}{Label of the plot. Default to \code{NULL}.}
  \item{xlab}{Label for the x-axis. Default to \code{NULL}.}
  \item{ylab}{Label for the y-axis. Default to \code{NULL}.}
}
\details{
This is an experimental function for the conjoint sequins, and thus might not
be fully utilized.
}
\value{
This function does not return anything.
}
\examples{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}