\name{plotCoverage}
\alias{plotCoverage}
\title{Plot sequence coverage}
\description{
Plot sequence coverage density for sequins.
}
\usage{
\method{plotCoverage}{TransQuin}(data, ylab='Sequence Coverage', seqs=NULL, minLen=NULL, maxLen=NULL)

\method{plotCoverage}{VarQuin}(data, ylab='Sequence Coverage', seqs=NULL, minLen=NULL, maxLen=NULL)

\method{plotCoverage}{FusQuin}(data, ylab='Sequence Coverage', seqs=NULL, minLen=NULL, maxLen=NULL)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{MetaQuin}, \code{LadQuin}, \code{StruQuin} or \code{FusQuin}. The data set needs to define a reference annotation and a bedgraph coverage.}
  \item{ylab}{Y-axis label of the plot}
  \item{seqs}{Filter sequins by names}
  \item{minLen}{Filter sequins by minimum length}
  \item{maxLen}{Filter sequins by maximum length}
}
\details{
Sequence coverage in sequins is an excellent proxy for the accompany genome. This function can be used to plot the density across the sequins. The sequins can be filtered by names and length.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
