\name{plotDensity}
\alias{plotDensity}
\title{Create density distribution}
\description{
Create density distribution for each sequin.
}
\usage{
\method{plotDensity}{TransQuin}(data)

\method{plotDensity}{VarQuin}(data)

\method{plotDensity}{FusQuin}(data)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin} or \code{FusQuin}. The data set needs to define a bedgraph file for the density.}
  \item{bedgr}{Bedgraph file for the distribution.}
  \item{annot}{Annotation file for the sequins.}  
}
\details{
The density (sequence depth) distribution of coverage for sequins is an excellent proxy for the accompany genomic. Users can use this function to tabulate the density across each of the sequins.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(VarQuin)
plotDensity(data)
}