\name{plotExpress}
\alias{plotExpress}
\title{Create an expression plot}
\description{
Create an expression plot where the x-axis is the input concentration and the y-axis is the measured abundance. This the most common but also most powerful data visualization in Anaquin.
}
\usage{
\method{plotExpress}{TransQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')

\method{plotExpress}{VarQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')

\method{plotExpress}{FusQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')

\method{plotExpress}{LadQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')

\method{plotExpress}{MetaQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have the input concentration and measured abundance.}
  \item{showStats}{Where to show the regression equation? \code{\link{'right'}} or \code{\link{'left'}}. }
  \item{title}{Label of the plot.}
  \item{xname}{Label for the x-axis}
  \item{yname}{Label for the x-axis}
  \item{showLOQ}{Show limit of quantification by piecewise-linear segmentation?}
}
\details{
Sequins are combined together across a range of concentrations to formulate a mixture. The mixture provides an internal scale against which to measure many features of an next-generation sequencing experiment, including the abundance of sequences in the accompying sample. The expression plot tabulates the input concentration against measured abundance, and evaluates the accuracy with a simple linear regression.

We would like to achieve a perfect correlation. However, this is typically not possible; instead we should use the regression to quantify the samples. The more corrected (slope and R2) the samples with the input concentration, the better experiment it is.

}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
# Plot the gene expression

data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotExpress(data)

# Plot the allele frequency

data(VarQuin)
data <- VarQuin(seqs=data$seqs,
                expected=data$expected,
                measured=data$measured)
plotExpress(data)
}