\name{plotFold}
\alias{plotFold}
\title{Create log-fold plot}
\description{
Create log-fold plot where the x-axis is the expected fold-change and the y-axis is the measured fold-change.
}
\usage{
\method{plotFold}{TransQuin}(data, title='TransQuin Differential', xlab='Expected log-fold (log2)', ylab='Measured log-fold (log2)')

\method{plotFold}{LadQuin}(data, title='', xlab='', ylab='')

\method{plotFold}{MetaQuin}(data, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have the expected fold-change and measured fold-change.}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
}
\details{
Scatter plot indicates correlation between observed and expected differential log-fold for each sequin gene. Each point represents a sequin, and they are colored based on the level of input concentration. The regression model is shown on the top-left corner. The blue line shows the regression line, and the shadow depicts the 95% confidence interval.

We would like to achieve a perfect correlation. However, this is typically not possible; the more corrected (slope and R2) the variables are, the better experiment it is.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotFold(data)
}