\name{plotFusion}
\alias{plotFusion}
\title{Create fusion expression plot}
\description{
Create a fusion expression plot where the x-axis is the logarithm of input concentration and the y-axis is the number of reads spanning the fusion breakpoints.
}
\usage{
\method{plotFusion}{FusQuin}(data, showLOQ=TRUE, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{FusQuin}. The data set needs to have the input concentration and measured abundance.}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
  \item{showLOQ}{Show limit of quantification by piecewise-linear segmentation?}
}
\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(FusQuin)
data <- FusQuin(seqs)
plotFusion(data)
}