\name{plotIExpress}
\alias{plotIExpress}
\title{Create isoform expression plot}
\description{
Create an isoform expression plot where the x-axis is the input concentration and the y-axis is the measured expression (eg: FPKM).
}
\usage{
\method{plotIExpress}{default}(data, title='Gene Expression', xlab='Input Concentrarion (log2)', ylab='FPKM (log2)')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin} or \code{MetaQuin}. The data set needs to have input concentration (\code{input}) and sensitivity (\code{measured}).}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
  \item{showLOQ}{Show limit of quantification?}
}
\details{
The x-axis gives the input concentration defined by the mixture on the logarithm scale. The y-axis gives the FPKM for each sequin, also on the logarithm scale. The limit-of-quantification (LOQ) (indicated by dashed line) is estimated by piecewise-linear segmentation. The regression model is shown on the top-left corner. The blue line shows the regression line, and the shadow depicts the 95 percent confidence interval.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(data)
plotIExpress(data)
}