\name{plotJunction}
\alias{plotJunction}
\title{Create a scatter plot with junction reads}
\description{
Create a scatter plot where the x-axis is the input concentration and the y-axis is the measured junction reads.
}
\usage{
\method{plotJunction}{FusQuin}(data, title=..., xlab=..., ylab=...)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{FusQuin}. The data set needs to have input concentration (\code{input}) and measured abundance (\code{measured}).}
  \item{title}{Label of the plot}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
}
\details{

}

\author{Ted Wong \email{t.wong@garvan.org.au}}