\name{plotKCounts}
\alias{plotKCounts}
\title{Create a k-mer coverage plot}
\description{
Create a k-mer coverage plot where the x-axis is the input concentration and the y-axis is the measured k-mer coverage.
}
\usage{
\method{plotKCounts}{MetaQuin}(data, showStats='left', showLOQ=TRUE, title='K-mer Detection', xlab='Input Concentration (log2 attomole/ul)', ylab='K-mer coverage (log2)')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have the input concentration and measured abundance.}
  \item{title}{Label of the plot.}
  \item{xname}{Label for the x-axis}
  \item{yname}{Label for the x-axis}
  \item{showLOQ}{Show limit of quantification?}
}
\details{
The mixture provides an internal scale against which to measure the k-mer abundance. This plot tabulates the input concentration against measured k-mer coverage, and evaluates the accuracy with a linear regression.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
