\name{plotLODR}
\alias{plotLODR}
\title{Create an LODR plot}
\description{
Create an Limit of Quantification Ratio (LODR) plot.
}
\usage{
\method{plotLODR}{TransQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}. The data set needs to have the measured abundance and empirical p-value.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotLODR(data)
}