\name{plotMA}
\alias{plotMA}
\title{Create MA plot}
\description{
Create an MA plot.
}
\usage{
\method{plotMA}{TransQuin}(data, title='', xname='', yname='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}. The data set needs to have average counts and log-ratios.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotMA(data)
}