\name{plotROC}
\alias{plotROC}
\title{Create an probability plot}
\description{
Create a probability plot.
}
\usage{
\method{plotProb}{TransQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotProb(data)
}