\name{plotReads}
\alias{plotReads}
\title{Create a scatter plot with read counts}
\description{
Create a scatter plot where the x-axis is the input concentration and the y-axis is the measured alignment read counts for each sequin.
}
\usage{
\method{plotReads}{TransQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{plotReads}{VarQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{plotReads}{FusQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{plotReads}{LadQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)

\method{plotReads}{MetaQuin}(data, showLOQ=TRUE, title=..., xlab=..., ylab=...)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have input concentration (\code{input}) and measured abundance (\code{measured}).}
  \item{title}{Label of the plot}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
  \item{showLOQ}{Show limit of quantification?}
}
\details{
Sequins are combined together across a range of concentrations to formulate a mixture. The mixture provides an internal scale against which to measure many features of an next-generation sequencing experiment, including alignment read counts in the accompying sample. The plot tabulates the input concentration against the read counts, and evaluates the accuracy with  simple linear regression.

We would like to achieve a perfect correlation. However, this is typically not possible; the more corrected (slope and R2) the variables are, the better experiment it is.

Limit-of-alignment (LOQ) is estimated if possible by piecewise linear segmentation.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=log2(data$input),
                  measured=log2(data$measured)
plotReads(data)
}