\name{plotROC}
\alias{plotROC}
\title{Create a ROC plot}
\description{
Create a receiver operating characteristic (ROC) plot. The curve is created by plotting the true positive rate (TPR) against the false positive rate (FPR) at various threshold settings.
}
\usage{
\method{plotROC}{TransQuin}(data, title='ROC Plot', color=NUL, guideTitle=NULL, showGuide=TRUE,  ...)

\method{plotROC}{VarQuin}(data, title='ROC Plot', color=NULL, guideTitle=NULL, showGuide=TRUE, ...)

\method{plotROC}{FusQuin}(data, title='ROC Plot', color=NULL, guideTitle=NULL, showGuide=TRUE,  ...)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin} or \code{FusQuin}}
  \item{title}{Title of the plot}
  \item{color}{Color of the curves, use \code{NULL} for default.}
  \item{guideTitle}{Title of the guide, use \code{NULL} for default.}
  \item{showGuide}{Show guides to the right?}
  \item{...}{For future expansion}
}
\author{Ted Wong \email{t.wong@garvan.org.au}}