\name{plotSensitivity}
\alias{plotSensitivity}
\title{Create sensitivity plot}
\description{
Create a sensitivity plot where the x-axis is input concentration and the y-axis is sensitivity. The plot should resemble a sigmoid function, where the sensitivity is zero for lower expressed sequins and one for highly expressed sequins.

The functions calculates limit of assembly (LOA).
}
\usage{
\method{plotSensitivity}{TransQuin}(data, title='', xlab='', ylab='', limit=0.98)

\method{plotSensitivity}{FusQuin}(data, title='', xlab='', ylab='', limit=0.98)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{FusQuin}. The data set needs to have the input concentration and measured abundance.}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the x-axis}
  \item{limit}{Limit for LOA}
}
\details{

}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
plotExpress(data)
}