\docType{data}
\name{RnaQuinMixtureB}
\alias{RnaQuinMixtureB}
\title{RnaQuin mixture B}
\format{Data frame:
\itemize{
  \item ID: Sequin name
  \item Length: Sequin length
  \item MixA: Input concentration
}}
\usage{
data(RnaQuinMixtureB)
}
\description{
Individual sequins are combined across a range of precise concentrations to formulate mixtures. By modulating the concentration at which each sequin is present in the mixture, we can emulate quantitative features of genome biology.

This is the mixture B in \code{RnaQuin}.
}
\value{
Data frame with columns defined in \code{Format}.
}