\docType{data}
\name{seqCuffcompare}
\alias{seqCuffcompare}
\title{Assembly sensitivity (RnaQuin)}
\format{Data frame:
\itemize{
  \item InputConcent: Input concentration in attomol/ul
  \item Sn: Measured sensitivity
}}
\usage{
data(seqCuffcompare)
}
\description{
Assembly sensitivity estimated by Cuffcompare from the RNA-Seq experiment of
Hardwick (2016). The data correspond to the first replicate of the samples.
}
\value{
A data frame with columns defined in \code{Format}.
}
\source{
S.A Hardwick. Spliced synthetic genes as internal controls in RNA sequencing
experiments. \code{Nature Methods}, 2016.
}