\name{TransDiff}
\alias{TransDiff}
\title{Quantify differential analysis}
\description{
Quantify differential analysis with TransQuin
}
\usage{
\method{TransDiff}{TransQuin}(data, r)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}.}
  \item{r}{Differential analysis object generated by \code{edgeR} or \code{DESeq2}.}
}
\details{
Refer to the user guide for details how to use edgeR and DESeq2 to work with this function.

Builds a linear model regressing the measured log-fold (dependent variable) to the input concentration (independent variable) defined by a mixture. We would like to achieve a perfect correlation. However, this is typically not possible; the more corrected (slope and R2) the samples with the input concentration, the better experiment it is.

This function also generates ROC curves and LODR curves.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotProb(data)
}