\docType{data}
\name{UserGuideData_5.4.5.1}
\alias{UserGuideData_5.4.5.1}
\title{Section 5.4.5.1 Assembly Dataset}
\format{Data frame:
\itemize{
  \item InputConcent: Input concentration in attomol/ul
  \item Sn: Measured sensitivity
}}
\usage{
data(UserGuideData_5.4.5.1)
}
\description{
Assembly sensitivity estimated by Cuffcompare. \code{Section 5.4.5.1}
of the Anaquin user guide has details on the data set.}
\value{
Data frame with columns defined in \code{Format}.
}
\source{
S.A Hardwick. Spliced synthetic genes as internal controls in RNA sequencing
experiments. \code{Nature Methods}, 2016.
}