\docType{data}
\name{Vignette_5.4.6.1}
\alias{Vignette_5.4.6.1}
\title{Quantify RnaQuin gene expression}
\format{A data frame:
\itemize{
  \item Length: Length of the sequin
  \item InputConcent: Input concentration in attomol/ul
  \item Sn: Measured sensitivity
}}
\usage{
data(Vignette_5.4.6.1)
head(Vignette_5.4.6.1)
}
\description{
A dataset containing assembly sensitivty estimated by Cufflinks. The vignette discusses the dataset 
in \bold{Section 5.4.6.1}.
}
\keyword{datasets}