\docType{data}
\name{Vignette_5.6.3}
\alias{Vignette_5.6.3}
\title{Quantify DESeq2 differential expression (RnaQuin)}
\format{A data frame:
\itemize{
  \item Length: Length of the sequin
  \item ExpLFC: Expected log-fold change 
  \item ObsLFC: Observed log-fold change
  \item SD: Standard deviation of the measurment
  \item Pval: P-value probability
  \item Qval: Q-value probability
  \item Mean: Average counts across the samples
}}
\usage{
data(Vignette_5.6.3)
head(Vignette_5.6.3)
}
\description{
A dataset containing differential gene expression estimated by DESeq2. The vignette discusses the dataset 
in \bold{Section 5.6.3}.
}
\keyword{datasets}