\docType{data}
\name{MixtureA}
\alias{MixtureA}
\title{RnaQuin mixture A}
\format{Data frame for each RnaQuin sequin:
\itemize{
  \item Length: Length in nucleotide bases
  \item MXA: Input concentration in attomol/ul
}}
\usage{
data(MixtureA)
}
\value{
A data frame with columns defined in \code{Format}.
}
\description{
RnaQuin sequins are combined together across a range of concentration to
formulate a mixture for emulating expression level. The concentration of
each sequin within a mixture is specificed in a CSV mixture file.

This is the staggered mixture (A) for RnaQuin, user can also download the
file directly from \code{http::/www.sequin.xyz}.
}