\name{plotAllele}
\alias{plotAllele}
\title{Create allele plot}
\description{
Create allele plot where the x-axis is the expected allele frequency and the y-axis is the measured allele frequency.
}
\usage{
\method{plotAllele}{default}(data, title=..., xlab=..., ylab=...)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}. The data set needs to have the expected allele frequency (\code{input}) and measured allele frequency (\code{measured}).}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
  \item{showLOQ}{Show limit of quantification?}
}
\details{
VarQuin are pairs of sequins that alternatively represent the reference and variant alleles with respect to the synthetic chromosome. By modulating the relative abundance of both the reference and the variant sequin, VarQuin enables a range of homozygous, heterozygous or heterogeneous allele frequencies to be emulated. For example, the use of only a single sequin can emulate a homozygous genotype, whilst the use of two sequins in equal parts can represent a heterozygous genotype.

The allele frequency plot tabulates the expected allele frequency against measured allele frequency, and evaluates the accuracy with a simple linear regression.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(VarQuin)
plotAllele(data)
}