\name{plotAssembly}
\alias{plotAssembly}
\title{Create assembly plot}
\description{
Create assembly plot where the x-axis is the logarithm of input concentration and the y-axis is the assembly sensitivity.
}
\usage{
\method{plotAssembly}{TransQuin}(data, title='Assembly Sensitivity',
                                 xlab='Input Concentration (log2 attomole/ul)',
                                 ylab='Sensitivity', limit=0.98)

\method{plotAssembly}{MetaQuin}(data,
                                title='Assembly Sensitivity',
                                xlab='Input Concentration (log2 attomole/ul)',
                                ylab='Sensitivity',
                                limit=0.98)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin} or \code{MetaQuin}. The data set needs to have input concentration (\code{input}) and sensitivity (\code{measured}).}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
  \item{showLOQ}{Show limit of assembly?}
}
\details{
The x-axis gives the input concentration defined by the mixture on the logarithm scale. The y-axis gives the sensitivity for each sequin, also on the logarithm scale. The colored points correspond to sequins, grouped by input concentration. The sigmoid function fitting is shown by a pink curve.

Limit-of-assemnly (LOA) is estimated if possible by fitting sigmoid function. It is the first sequin with a sensitivity of \code{limit}.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
plotAssembly(data)

data(MetaQuin)
plotAssembly(data)
}