\name{plotDensity}
\alias{plotDensity}
\title{Create density distribution}
\description{
Create density distribution for each sequin.
}
\usage{
\method{plotDensity}{default}(data)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}. The data set needs to define a bedgraph file for the density. It is recommended to use the command-line \code{VarCoverage} to auto-generate a script for this function.}
}
\details{
The density (sequence depth) distribution of coverage for sequins is an excellent proxy for the accompany genomic. Users can use this function to tabulate the density across each of the sequins.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(VarQuin)
plotDensity(data)
}