\name{plotExpress}
\alias{plotExpress}
\title{Create a fold plot}
\description{
Create a fold plot where the x-axis is the expected fold-change and the y-axis is the measured fold-change.
}
\usage{
\method{plotFold}{TransQuin}(data, showStats='left', title='', xlab='', ylab='')

\method{plotFold}{FusQuin}(data, showStats='left', title='', xlab='', ylab='')

\method{plotFold}{LadQuin}(data, showStats='left', title='', xlab='', ylab='')

\method{plotFold}{MetaQuin}(data, showStats='left', title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{FusQuin}, \code{MetaQuin} or \code{LadQuin}. The data set needs to have the expected fold-change and measured fold-change.}
  \item{showStats}{Where to show the regression equation? \code{\link{'right'}} or \code{\link{'left'}}. }
  \item{title}{Label of the plot.}
  \item{xname}{Label for the x-axis}
  \item{yname}{Label for the x-axis}
}
\details{
Sequins are combined together across a range of concentrations to formulate a mixture. The mixture provides an internal scale against which to measure many features of an next-generation sequencing experiment, including the fold-change in the accompying sample.

This function requires two input mixtures, which combined to form a scaling of fold-change. The fold plot tabulates the fold-change of the input concentrations against measured fold-change, and evaluates the accuracy with a simple linear regression.

We would like to achieve a perfect correlation. However, this is typically not possible; instead we should use the regression to quantify the samples. The more corrected (slope and R2) the samples with the expected fold-change, the better experiment it is.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
# Plot differential analysis (expected log-fold against the measured log-fold)

data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$eLFold,
                  measured=data$mLFold)
plotFold(data)
}