\name{plotLinear}
\alias{plotLinear}
\title{Plot linear model for sequins}
\description{
Create a scatter plot with expected abundance on the x-axis, and measured
abundance on the y-axis.
}
\usage{
plotLinear(data, title, xlab, ylab, showSD, showLOQ, xBreaks, yBreaks, errors, showAxis, ...)
}
\arguments{
  \item{data}{Anaquin dataset created by \code{AnaquinData}. It needs to define
              information in \bold{Details}.}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
  \item{xBreaks}{Breaks for the x-axis}
  \item{yBreaks}{Breaks for the y-axis}
  \item{showSD}{Show standard deviation bars vertically? Default to
                \code{TRUE}.}
  \item{showLOQ}{Show limit-of-quantification? Default to \code{TRUE}.}
  \item{errors}{How errors bar should be calculated. \code{SD} or \code{Range}}
  \item{showAxis}{Show x-axis and y-axis? Default to \code{TRUE}.}
  \item{...}{Reserved for internal testing}  
}
\value{
The functions does not return anything but it prints a scatter plot.
}
\details{

\code{plotLinear} requires the following data inputs from \code{AnaquinData}.

\tabular{rlll}{
    \tab \code{seqs}   \tab List of sequin identifiers (eg. R2_11_2).\cr
    \tab \code{input}   \tab Input concentration of sequins in attomol/ul.\cr
    \tab \code{measured} \tab Measured abundance of sequins (typically FPKM).\cr
}

The \code{plotLinear} function plots a scatter plot with expected abundance on
the x-axis, and measured abundance on the y-axis. The expected abundance is
typically the input concentration of sequins in the mixture, although other
measures (such as expected allele frequency) are also possible. The function
builds a linear regression between the two variables, and reports associated
statistics (R2, correlation and regression parameters) on the plot.

The function also estimates limit-of-quantification (LOQ) breakpoint, and
reports it on the plot if found. LOQ is defined as the lowest empirical
detection limit, a threshold value beyond which stochastic behavior occur.
LOQ is estimated by fitting segmented linear regression with two segments
on the entire data set, while minimizing the total sum of squares of the
differences between the variables. 
}

\examples{
library(Anaquin)

#
# Data set generated by Cufflinks and Anaquin. described in Section 5.4.6.3 of
# the user guide.
#
data(UserGuideData_5.4.6.3)

title <- 'Gene Expression'
xlab  <- 'Input Concentration (log2)'
ylab  <- 'FPKM (log2)'

# Sequin names
seqs <- row.names(UserGuideData_5.4.6.3)

# Input concentration
input <- log2(UserGuideData_5.4.6.3$InputConcent)

# Measured FPKM
measured <- log2(UserGuideData_5.4.6.3[,2:4])

anaquin <- AnaquinData(analysis='PlotLinear',
                           seqs=seqs,
                          input=input,
                       measured=measured)

plotLinear(anaquin, title=title, xlab=xlab, ylab=ylab, showLOQ=TRUE)
}

\author{Ted Wong \email{t.wong@garvan.org.au}}