\name{plotLODR}
\alias{plotLODR}
\title{Create Limit-of-Detection Ratio (LODR) plot}
\description{
Create a Limit-of-Detection Ratio (LODR) plot between measured abundance
(x-axis) and p-value probability (y-axis).
}
\usage{
plotLODR(data, FDR, title, xlab, ylab, legTitle, showConf, ...)
}
\arguments{
  \item{data}{Anaquin dataset created by \code{AnaquinData}. It needs to
              define information in \bold{Details}.}
  \item{FDR}{Chosen false-discovery-rate. Default to \code{NULL}.}
  \item{title}{Label of the plot. Default to \code{NULL}.}
  \item{xlab}{Label for the x-axis. Default to \code{NULL}.}
  \item{ylab}{Label for the y-axis. Default to \code{NULL}.}
  \item{legTitle}{Title for the legend. Default to \code{'Ratio'}.}
  \item{showConf}{Show confidence interval? Default to \code{FALSE}.}
  \item{...}{Reserved for internal testing}
}
\details{

\code{plotLODR} requires the following data inputs from \code{AnaquinData}.

\tabular{rlll}{
    \tab \code{seqs}    \tab List of sequin identifiers (eg. R2_11_2).\cr
    \tab \code{measured} \tab Measured abundance (eg: average counts, DP field
                              in a VCF file etc)\cr
    \tab \code{ratio}    \tab Expected ratio; eg: expected log-fold ratio or
                              expected allele frequency etc\cr
    \tab \code{pval}     \tab P-value probability\cr
}

Create a Limit-of-Detection Ratio (LODR) plot between measured abundance
(x-axis) and p-value probability (y-axis).

The LODR plot indicates the confidence in measurement relative to the magnitude
of the measurement. For example, p-value should converge to zero as the
sequencing depth increases.

The function also fits non-parametric curves for each sequin ratio group. The
curves are modelled with local regression analysis, and are colored by the
sequin group.

plotLODR is an amendment from the LODR code in the ERCC dashboard R-package.
Further details on the statistical algorithm is available in the ERCC
documentation at
https://bioconductor.org/packages/release/bioc/html/erccdashboard.html.
}

\value{
The functions does not return anything but it prints a LODR plot.
}

\examples{
library(Anaquin)

#
# Data set generated by DESeq2 and Anaquin. described in Section 5.6.3.3 of
# the user guide.
#
data(UserGuideData_5.6.3)

xlab  <- 'Average Counts'
ylab  <- 'P-value'
title <- 'LODR Curves'

# Sequin names
seqs <- row.names(UserGuideData_5.6.3)

# Expected log-fold
ratio <- UserGuideData_5.6.3$ExpLFC

# Measured average abundance
measured <- UserGuideData_5.6.3$Mean

# P-value
pval <- UserGuideData_5.6.3$Pval

# Q-value
qval <- UserGuideData_5.6.3$Qval

anaquin <- AnaquinData(analysis='PlotLODR',
                           seqs=seqs,
                      measured=measured,
                         ratio=ratio,
                          pval=pval,
                          qval=qval)

plotLODR(anaquin, xlab=xlab, ylab=ylab, title=title, FDR=0.1)
}

\author{Ted Wong \email{t.wong@garvan.org.au}}