\name{plotLODR}
\alias{plotLODR}
\title{Create LODR plot}
\description{
Create an Limit of Quantification Ratio (LODR) plot.
}
\usage{
\method{plotLODR}{TransQuin}(data, title='', xname='', yname='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}. The data set needs to have the measured abundance and empirical p-value.}
    \item{chosenFDR}{The chosen false-discovery rate threshold.}
}
\details{

Quote from: Munro 2014:

------------------------------------------------------------------------------

\emph{
Identifying differentially expressed transcripts is the objective of differential expression experiments; however, how much information (signal) is needed to have confidence that a given fold change in expression of transcripts will be detected? With LODR estimates, empirical ERCC control ratio measurements can inform researchers of diagnostic power at a given fold change relative to transcript abundance for an experiment.

An LODR estimate for a particular fold change is the minimum signal, above which differentially expressed transcripts can be detected with a specified degree of confidence. LODR offers a statistically derived, objective alternative to other methods of parsing gene lists.

An LODR estimate is obtained for a specified ratio by modelling the relationship between differential expression test P-values and signal. An acceptable false discovery rate (FDR) must be chosen to estimate an LODR. For the selected FDR (q-value) a threshold P-value can be selected from the population of P-values from the experiment. An LODR estimate for each differential ratio is found based on the intersection of the model confidence interval upper bound with the P-value threshold.
}

------------------------------------------------------------------------------
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotLODR(data, chosenFDR=0.1)
}