\name{plotMA}
\alias{plotMA}
\title{Create an MA plot}
\description{
Create an MA plot.
}
\usage{
\method{plotMA}{TransQuin}(data)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotLODR(data)
}