\name{plotROC}
\alias{plotROC}
\title{Create a p-value probability plot}
\description{
Create a p-value probability plot where the x-axis is the average counts and the y-axis it the p-value probability.
}
\usage{
\method{plotProb}{VarQuin}(data, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}. It needs to have average counts and p-value \code{pval} defined. }
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
}
\details{
The plot is useful for depicting the relationship between the abundance (average counts for repliacte) and effect size (p-value).
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(VarQuin)
data <- VarQuin(seqs)
plotProb(data)
}