\name{plotProb}
\alias{plotProb}
\title{Create a probability plot}
\description{
Create a probability plot.
}
\usage{
\method{plotMajor}{TransQuin}(data, showStats='left', showLOQ=TRUE, title='', xlab='', ylab='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotLODR(data)
}