\name{plotROC}
\alias{plotROC}
\title{Create an ROC plot}
\description{
Create an receiver operating characteristic (ROC) plot.
}
\usage{
\method{plotROC}{TransQuin}(data, title='')

\method{plotROC}{VarQuin}(data, title='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin} or \code{VarQuin}.}
  \item{title}{Title of the ROC plot.}  
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotROC(data)
}