\name{plotROC}
\alias{plotROC}
\title{Create an ROC plot}
\description{
Create an receiver operating characteristic (ROC) plot.
}
\usage{
\method{plotROC}{TransQuin}(data, title='')

\method{plotROC}{VarQuin}(data, title='')

\method{plotROC}{FusQuin}(data, title='')

\method{plotROC}{MetaQuin}(data, title='')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin} or \code{MetaQuin}.}
  \item{title}{Title of the ROC plot.}  
}
\details{
\code{TransQuin}

The plot gives the ability to classify sequins based on their expected fold-change relative to a reference group. A false positive is a sequin that is indistinguishable from the reference group. Similarly, a true positive is a sequin that is distingushable from the reference group. The sequins with fold-change of 1 are chosen as the reference group.

\code{VarQuin}

\code{FusQuin}

\code{MetaQuin}

}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotROC(data)
}