\name{plotROC}
\alias{plotROC}
\title{Create an ROC plot}
\description{
Create an receiver operating characteristic (ROC) plot, where the x-axis is false-positives (FP) and the y-axis is true-positivies (TP).
}
\usage{
\method{plotROC}{TransQuin}(data, title='', xlab='TP', ylab='FP')

\method{plotROC}{VarQuin}(data, title='', xlab='TP', ylab='FP')

\method{plotROC}{FusQuin}(data, title='', xlab='TP', ylab='FP')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}, \code{VarQuin}, \code{FusQuin} or \code{MetaQuin}.}
  \item{title}{Title of the ROC plot.}
  \item{xlab}{Label for the x-axis.}
  \item{ylab}{Label for the y-axis}
  \item{showAUC}{Show AUC?}
  \item{showGuide}{Show guides to thr right?}
}
\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotROC(data)
}