\name{plotSplice}
\alias{plotSplice}
\title{Create an alternative splicing plot}
\description{
Create an alternative splicing plot where the x-axis is expected minor isoform fraction and the y-axis is the measured fraction. Both axis should be on the logarithm scale.
}
\usage{
\method{plotSplice}{TransQuin}(data, showStats='left', showLOQ=TRUE, title='Minor Fraction', xlab='Expected minor isoform fraction (log 2)', ylab='Measured minor isoform fraction (log2)')
}
\arguments{
  \item{data}{Anaquin data set generated by \code{TransQuin}.}
  \item{title}{Label of the plot.}
  \item{xlab}{Label for the x-axis}
  \item{ylab}{Label for the y-axis}
}
\details{
Scatter plot indicates correlation between observed and expected minimum isoform as a fraction of the major isoform for each sequin gene. Each point represents a sequin, and they are colored based on the level of input concentration.
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs)
plotSplice(data)
}