\name{RnaQuin.genes}
\alias{RnaQuin.genes}
\alias{RnaQuin.genes,AnaquinData-method}
\title{Return RnaQuin genes}
\description{
Return RnaQuin genes from data set created by \code{AnaquinData}.
}
\usage{
RnaQuin.genes(object, ...)
\S4method{RnaQuin.genes}{AnaquinData}(object)
}
\arguments{
  \item{object}{S4 object created by \code{AnaquinData}}
  \item{...}{Not used}
}
\details{
This function takes an \code{AnaquinData} object and return the RnaQuin
genes identifiers.
}
\value{
Data frame with the genes and their input concentration. Return \code{NULL}
if unavailable.
}
\examples{
library(Anaquin)

#
# Data set generated by Cufflinks and Anaquin. described in Section 5.4.5.1 of
# the user guide.
#
data(UserGuideData_5.4.5.1)

data <- UserGuideData_5.4.5.1

# Sequin names
seqs <- row.names(data)

# Input concentration
input <- log2(data$InputConcent)

anaquin <- new('AnaquinData', analysis = 'Mixture',
                                  seqs = row.names(data),
                                 input = log2(data$InputConcent))
RnaQuin.genes(anaquin)
}

\author{Ted Wong \email{t.wong@garvan.org.au}}