\name{TransDiff}
\alias{TransDiff}
\title{Quantify differential analysis}
\description{
Quantify differential analysis with TransQuin
}
\usage{
\method{TransKDiff}{TransQuin}(data)
}
\arguments{
  \item{data}{Anaquin data set generated by \code{VarQuin}.}
}
\details{
}

\author{Ted Wong \email{t.wong@garvan.org.au}}
\examples{
data(TransQuin)
data <- TransQuin(seqs=data$seqs,
                  expected=data$expected,
                  measured=data$measured)
plotProb(data)
}